+++ src/arm.rs
@@ -1,4 +1,4 @@
-use bevy::prelude::{Bundle, Capsule3d, Component, Entity};
+use bevy::prelude::{Bundle, Capsule3d, Component, Entity, Transform};
 use bevy_rapier3d::parry::math::Real;
 use bevy_rapier3d::prelude::{Collider, RigidBody};
 
@@ -32,6 +32,7 @@ impl BodySegment for CapsuleSegment {
 pub struct CapsuleSegmentBundle {
     rb: RigidBody,
     collider: Collider,
+    transform: Transform,
     capsule_segment: CapsuleSegment,
 }
 
@@ -39,6 +40,7 @@ impl CapsuleSegmentBundle {
     pub fn new(capsule_info: Capsule3d) -> Self {
         return Self {
             rb: RigidBody::Dynamic,
+            transform: Transform::default(),
             collider: Collider::capsule_y(capsule_info.half_length, capsule_info.radius),
             capsule_segment: CapsuleSegment {
                 capsule: capsule_info
