+++ src/ik.rs
@@ -1,8 +1,8 @@
 use bevy::prelude::{App, Bundle, Component, Entity, Last, Plugin};
 use bevy::utils::default;
-use k::{Chain, Error, InverseKinematicsSolver, Isometry3, JacobianIkSolver, NodeBuilder, RealField, SerialChain, SubsetOf};
+use k::{Chain, connect, Error, InverseKinematicsSolver, Isometry3, JacobianIkSolver, JointType, NodeBuilder, RealField, SerialChain, SubsetOf};
 use crate::arm::{ArmInfo, BodySegment, CapsuleSegment};
-use crate::ik_systems::initialize_ik_arms;
+use crate::ik_systems::set_ik_arm_positions;
 
 
 pub struct IKPlugin;
@@ -13,13 +13,10 @@ impl Plugin for IKPlugin {
 
         app
             .register_component_as::<dyn BodySegment, CapsuleSegment>()
-            .add_systems(Last, initialize_ik_arms);
+            .add_systems(Last, set_ik_arm_positions);
     }
 }
 
-
-#[derive(Component)]
-pub struct UninitializedIKArm;
 
 #[derive(Component)]
 pub struct IKArm<T: RealField> {
@@ -41,23 +38,59 @@ where
     T: RealField + SubsetOf<f64>
 {
     fn default() -> Self {
+        let fixed: k::Node<T> = NodeBuilder::new()
+            .name("fixed")
+            .joint_type(JointType::Fixed)
+            .finalize()
+            .into();
+        let l0: k::Node<T> = NodeBuilder::new()
+            .name("shoulder_x_rot")
+            .joint_type(JointType::Rotational {
+                axis: k::Vector3::x_axis()
+            })
+            .finalize()
+            .into();
+        let l1: k::Node<T> = NodeBuilder::new()
+            .name("shoulder_y_rot")
+            .joint_type(JointType::Rotational {
+                axis: k::Vector3::y_axis()
+            })
+            .finalize()
+            .into();
+        let l2: k::Node<T> = NodeBuilder::new()
+            .name("shoulder_z_rot")
+            .joint_type(JointType::Rotational {
+                axis: k::Vector3::z_axis()
+            })
+            .finalize()
+            .into();
+        let l3: k::Node<T> = NodeBuilder::new()
+            .name("elbow_y_rot")
+            .joint_type(JointType::Rotational {
+                axis: k::Vector3::y_axis()
+            })
+            .finalize()
+            .into();
+        let l4: k::Node<T> = NodeBuilder::new()
+            .name("wrist")
+            .finalize()
+            .into();
+        connect![fixed => l0 => l1 => l2 => l3 => l4];
+
         Self {
-            arm_chain: SerialChain::new_unchecked(Chain::from_root(
-                NodeBuilder::new().finalize().into()
-            )),
+            arm_chain: SerialChain::new_unchecked(k::Chain::from_root(fixed)),
             ik_solver: JacobianIkSolver::default()
         }
     }
 }
 
 #[derive(Bundle)]
-pub struct UninitIKArmBundle<T: RealField> {
+pub struct IKArmBundle<T: RealField> {
     pub arm_info: ArmInfo,
     pub ik_arm: IKArm<T>,
-    pub _uninit: UninitializedIKArm
 }
 
-impl<T> UninitIKArmBundle<T> 
+impl<T> IKArmBundle<T>
 where 
     T: RealField + SubsetOf<f64>
 {
@@ -68,20 +101,18 @@ where
                 lower_arm
             },
             ik_arm: default(),
-            _uninit: UninitializedIKArm,
         }
     }
 }
 
-impl<T> Default for UninitIKArmBundle<T>
+impl<T> Default for IKArmBundle<T>
 where
     T: RealField + SubsetOf<f64>
 {
     fn default() -> Self {
         Self {
-            arm_info: ArmInfo::default(),
-            ik_arm: IKArm::default(),
-            _uninit: UninitializedIKArm
+            arm_info: default(),
+            ik_arm: default(),
         }
     }
 }
